re rename migration & requested changes
